Adds support for resin.io
Adds a dockerfile and resin build script.

Note that due to a bug in resin.io we cannot run `cargo build` on
resin.io (it segfaults as it is actually running under qemu which does
not like multiple threads). Instead we build locally, create a tempory
repo and upload that to resin.io with everything prebuilt - these steps
are in the `resin` script.

Also tweaks some settings to work with the public/url forwarding:

* Changes port from 3000 to 80 as resin.io only supports port 80.
* Changes the timeouts and frequency of the commands as we are now going
over the internet.